add decimalField in AllTables

+++ EnergyCalculator.Core/Contracts/IIngredientService.cs
@@ -22,6 +22,6 @@ namespace EnergyCalculator.Core.Contracts
         Task<bool> ExistsIngredient(int productId, int receiptId);
 
         Task<IEnumerable<AllIngredientViewModel>> MyIngredients(string userId);
-        Task<int> CalculateTotalCaloriesByProduct(int productId, int totalForIngredient);
+        Task<double> CalculateTotalCaloriesByProduct(double productId, double totalForIngredient);
     }
 }
