+++ EnergyCalculator.Infrastructure/Data/Entities/Receipt.cs
@@ -16,7 +16,7 @@ namespace EnergyCalculator.Infrastructure.Data.Entities
         [StringLength(30)]
         public string Name { get; set; } = null!;
 
-       public int TotalCalories { get; set; }
+       public double TotalCalories { get; set; }
 
          public string UserId { get; set; }= null!;
